changed to versioning scheme to 2.1.2.0
changed the versioning scheme to more closely follow changes to the pusher lib.  The last 0 is our patch number since major changes to this code are ONLY introduced with changes to the pusher lib.  Now it will be more clear which version of pusher is supported by any given release.